+++ rust-runtime/aws-smithy-http-server/src/lib.rs
@@ -28,13 +28,10 @@ pub mod routing;
 #[doc(hidden)]
 pub mod runtime_error;
 
-#[doc(hidden)]
-pub mod routers;
-
 #[doc(inline)]
 pub(crate) use self::error::Error;
-pub use self::request::extension::Extension;
 #[doc(inline)]
+pub use self::request::extension::Extension;
 #[doc(inline)]
 pub use tower_http::add_extension::{AddExtension, AddExtensionLayer};
