+++ rust-runtime/aws-smithy-http-server/src/proto/rest_json_1/router.rs
@@ -7,7 +7,7 @@ use crate::body::BoxBody;
 use crate::extension::RuntimeErrorExtension;
 use crate::proto::rest::router::Error;
 use crate::response::IntoResponse;
-use crate::routers::{method_disallowed, UNKNOWN_OPERATION_EXCEPTION};
+use crate::routing::{method_disallowed, UNKNOWN_OPERATION_EXCEPTION};
 
 use super::RestJson1;
